feat(test) tested successfully 51 test on 'StrategyRecursiveFarming' contract, and divided all the tests in 5 different files. added 'claimRewards' function tests with its mocks too
